Fully implemented Route53 internal DNS for VPCs   - EC2, RDS, and Redis Replication instances all get automatically     added to DNS   - Boss-utils firstboot script configures the DNS search path so that     non-qualified DNS names will resolve (vault vs vault.production.boss)   - Moved from dynamically creating the AWS section of boss.config to     using static entries because DNS handles the resolution   - Removed any reference to hardcoded IP addresses for EC2 instances     as RDS and Redis didn't have specific IPs and DNS allows us to not     care about the specific IP
The Core configuration now creates a subnet for each availability zone
in the region that is being used to create the configuration. The subnets
are named with the one character leter for the availability zone
  - Both production and proofreader now make use of these
  - There is a config object method for looking up all of these subnets

The Production configuration creates two Redis cache instances. A single
node ElastiCache Redis Cache Cluster is not used because CloudFormation
cannot resolve the endpoint address dynamically (to create the DNS record)
so a single node ElastiCache Redis Replication Group is used instead.
  - The smallest node type of cache.m3.medium is used, as cache.t2.* cannot
    be used for ElastiCache Replication Groups.

This is also setting us up to handle failover and autoscale groups
with minimal interruption
  - Route53 has some built in failover capability
  - Relying on DNS instead of IP addresses decouples components