Update update_lamba_fcn.py - cherrypicked 90d4e7.
Put write_s3_index_lambda.py in the root of zip.

Put this lambda there so it can be imported by lambda_loader.py.  This
lambda returns a value so it can't use runpy.run_path().  Because it
lives in a lambda folder, it can't be imported from that location.
Thus, place it in the root of the zip.

Include heaviside in lambda build.