Add s3:ListBucket permission to cuboid import role
Let the cuboid import lambda list the contents of a bucket so it can
handle errors properly when the cuboid was deleted from the bucket.
Without this permission, it would get an Access Denied Error when trying
to get the cuboid's metadata or invoke the copy operation.

The cuboid could be deleted by another invocation of the lambda.

Updated add_s3_bucket_policy() and append_s3_bucket_policy() so that
actions such as s3:ListBucket can be specified (must only specify the
bucket, not <bucket>/* because it operates on the bucket, not the
bucket's objects).